Auto-split asset on withdraw on asset level

+++ app/src/main/java/com/ceaver/assin/positions/Position.kt
@@ -6,7 +6,7 @@ import java.math.MathContext
 import java.time.LocalDate
 
 data class Position(
-        val id: Int,
+        val id: BigDecimal,
         val title: Title,
         val amount: BigDecimal,
         val openDate: LocalDate,
